checkpoint save: moving interpolate and resample functions out of TSDF
