+++ python/tempo/utils.py
@@ -2,18 +2,14 @@ from __future__ import annotations
 
 import logging
 import os
-import warnings
-from typing import List, Optional, Union, overload
+from typing import Optional, Union, overload
 
 from IPython import get_ipython
 from IPython.core.display import HTML
 from IPython.display import display as ipydisplay
 from pandas.core.frame import DataFrame as pandasDataFrame
-
-import pyspark.sql.functions as sfn
 from pyspark.sql.dataframe import DataFrame
 
-import tempo.resample as t_resample
 import tempo.tsdf as t_tsdf
 
 logger = logging.getLogger(__name__)
@@ -25,102 +21,6 @@ DB_HOME env variable has been chosen and that's because this variable is a speci
 This constant is to ensure the correct behaviour of the show and display methods are called based on the platform
 where the code is running from.
 """
-
-
-class ResampleWarning(Warning):
-    """
-    This class is a warning that is raised when the interpolate or resample with fill methods are called.
-    """
-
-    pass
-
-
-def calculate_time_horizon(
-    tsdf: t_tsdf.TSDF,
-    freq: str,
-    local_freq_dict: Optional[t_resample.FreqDict] = None,
-) -> None:
-    # Convert Frequency using resample dictionary
-    if local_freq_dict is None:
-        local_freq_dict = t_resample.freq_dict
-    parsed_freq = t_resample.checkAllowableFreq(freq)
-    period, unit = parsed_freq[0], parsed_freq[1]
-    if t_resample.is_valid_allowed_freq_keys(
-        unit,
-        t_resample.ALLOWED_FREQ_KEYS,
-    ):
-        freq = f"{period} {local_freq_dict[unit]}"  # type: ignore[literal-required]
-    else:
-        raise ValueError(f"Frequency {unit} not supported")
-
-    # Get max and min timestamp per partition
-    if tsdf.series_ids:
-        grouped_df = tsdf.df.groupBy(*tsdf.series_ids)
-    else:
-        grouped_df = tsdf.df.groupBy()
-    ts_range_per_series: DataFrame = grouped_df.agg(
-        sfn.max(tsdf.ts_col).alias("max_ts"),
-        sfn.min(tsdf.ts_col).alias("min_ts"),
-    )
-
-    # Generate upscale metrics
-    normalized_time_df: DataFrame = (
-        ts_range_per_series.withColumn("min_epoch_ms", sfn.expr("unix_millis(min_ts)"))
-        .withColumn("max_epoch_ms", sfn.expr("unix_millis(max_ts)"))
-        .withColumn(
-            "interval_ms",
-            sfn.expr(
-                f"unix_millis(cast('1970-01-01 00:00:00.000+0000' as TIMESTAMP) + INTERVAL {freq})"
-            ),
-        )
-        .withColumn(
-            "rounded_min_epoch",
-            sfn.expr("min_epoch_ms - (min_epoch_ms % interval_ms)"),
-        )
-        .withColumn(
-            "rounded_max_epoch",
-            sfn.expr("max_epoch_ms - (max_epoch_ms % interval_ms)"),
-        )
-        .withColumn("diff_ms", sfn.expr("rounded_max_epoch - rounded_min_epoch"))
-        .withColumn("num_values", sfn.expr("(diff_ms/interval_ms) +1"))
-    )
-
-    (
-        min_ts,
-        max_ts,
-        min_value_partition,
-        max_value_partition,
-        p25_value_partition,
-        p50_value_partition,
-        p75_value_partition,
-        total_values,
-    ) = normalized_time_df.select(
-        sfn.min("min_ts"),
-        sfn.max("max_ts"),
-        sfn.min("num_values"),
-        sfn.max("num_values"),
-        sfn.percentile_approx("num_values", 0.25),
-        sfn.percentile_approx("num_values", 0.5),
-        sfn.percentile_approx("num_values", 0.75),
-        sfn.sum("num_values"),
-    ).first()
-
-    warnings.simplefilter("always", ResampleWarning)
-    warnings.warn(
-        f"""
-            Resample Metrics Warning:
-                Earliest Timestamp: {min_ts}
-                Latest Timestamp: {max_ts}
-                No. of Unique Partitions: {normalized_time_df.count()}
-                Resampled Min No. Values in Single a Partition: {min_value_partition}
-                Resampled Max No. Values in Single a Partition: {max_value_partition}
-                Resampled P25 No. Values in Single a Partition: {p25_value_partition}
-                Resampled P50 No. Values in Single a Partition: {p50_value_partition}
-                Resampled P75 No. Values in Single a Partition: {p75_value_partition}
-                Resampled Total No. Values Across All Partitions: {total_values}
-        """,
-        ResampleWarning,
-    )
 
 
 def _is_capable_of_html_rendering() -> bool:
